scrapers/requirements.txt
crawl4ai>=0.8.0
asyncpg>=0.30.0
pydantic>=2.10.0
anthropic>=0.42.0
requests>=2.32.0
python-dotenv>=1.0.0
